+++ simple_history/management/commands/populate_history.py
@@ -1,4 +1,3 @@
-import django
 from django.apps import apps
 from django.core.management.base import BaseCommand, CommandError
 
@@ -127,9 +126,7 @@ class Command(BaseCommand):
                 )
             )
 
-        iterator_kwargs = (
-            {"chunk_size": batch_size} if django.VERSION >= (2, 0, 0) else {}
-        )
+        iterator_kwargs = {"chunk_size": batch_size}
         for index, instance in enumerate(
             model._default_manager.iterator(**iterator_kwargs)
         ):