+++ simple_history/utils.py
@@ -140,11 +140,6 @@ def bulk_update_with_history(
     :param manager: Optional model manager to use for the model instead of the default
         manager
     """
-    if django.VERSION < (2, 2,):
-        raise NotImplementedError(
-            "bulk_update_with_history is only available on "
-            "Django versions 2.2 and later"
-        )
     history_manager = get_history_manager_for_model(model)
     model_manager = manager or model._default_manager
     if model_manager.model is not model: